Fix mismatched image and tags in convert script
The share/zip strategy didn't guarantee that different asynchronous
branches/pipes would get pulled back together in the same order -- we
were finding that images were tagged or attributed with details from
other images. It took a while to understand the forkJoin api, but it
seems to be more reliable as it's forcing all of the related work into
the same scope/closure.