Create commands for converting array of objects
We're using R and the jsonlite package to generate a lot of
our json files. The package makes it really easy to
convert a dataframe to an array of json objects, but
for our import function we generally either want a
folder of individual files, or one large associative array.
The new split method will take that array of objects and
write each one to a separate file in a folder with the id as
the file name. The associate method will similiary create
an object with one key/value pair for each of the objects
in the array -- with the key taking the value of the id
from the nested object.