Try out options for incremental exports
We don't want to export every document in a collection every time we do
an export. First, it uses up more of our read quota and second, until we
have a canonical format, the same object exported twice could end up
looking different in the json output just because of the relatively
random order that the fields get written out.

It would be nice if we could commit out exports to a repository like
git to make roll backs to various points in time easier. It would also
be nice if it was easy to see from those commits what had changed over
time and when. Canonicalizing and "prettifying" the output would go a
long way to make that easier, but incremental exports would also help.

I ran a quick proof of concept to see if we can export based on the
updated date field (it will only work as long as the date is
consistently updated by all methods that change the documents). It
worked, but I need to build out the method for saving the max timestamp
from the previous export for use in the next export -- a so called
highwater mark. For now I've just commented out the where clause as a
bit of documentation to help with the actual implementation later.