Refactor build and publish to github packages
I'm doing a bit of refactoring in order to publish only the necessary
files to the github node registry. We don't want to publish all of the
source files -- just the CLI script, the typescript headers and the
transpiled javascript. After refactoring the build, all of that now gets
written out to the dist folder (formerly the lib folder) and I've
configured the package publishing to whitelist all of the files in the
dist folder (which automatically ignores everything else except
fortunately the bash script file itself).